pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "conviction-tracker"
version = "0.1.0"
description = "Insider & Congressional Trade Signal System"
requires-python = ">=3.11"
dependencies = [
    "edgartools>=2.0",
    "streamlit>=1.32",
    "plotly>=5.20",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
